feat: add link to principles from main page, improve layout on principles page
