+++ src/pages/index.tsx
@@ -50,7 +50,9 @@ const Home: NextPage = () => {
             kontinuitet, gode insentiver og Super Mario-effekten er noe av det
             vi mener kreves for å lykkes.
           </Paragraph>
-          <Link sx={{ color: 'black100' }}>Les om våre prinsipper</Link>
+          <Link sx={{ color: 'black100' }} href={'/principles'}>
+            Les om våre prinsipper
+          </Link>
         </Card>
         <Grid
           columns={['repeat(3, 1fr)']}
